# docs/openapi.yml
openapi: 3.0.0
servers:
  - description: Production
    url: http://localhost:9090
info:
  description: Codewind API
  version: 0.1.0
  title: Codewind API
  contact:
    name: API support
    email: example@ibm.com
  license:
    name: Microclimate Developer License
    url: 'https://github.ibm.com/dev-ex/codewind/blob/master/LICENSE/LICENSE.md'

# tags:
  # - name: users
  #   description: Operations available to regular users
  # - name: admin
  #   description: Operations available only to authenticated admins
paths:
...
  /api/v1/projects/{id}/capabilities:
    get:
      summary: Get a list of supported capabilities for the project
      parameters:
          - in: path
            name: id
            schema:
              type: string
            required: true
            description: id of project
      responses:
        200:
          description: Successfully completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  startModes:
                    type: array
                    items:
                      type: string
                      enum: ['run', 'debug']
                  controlCommands:
                    type: array
                    items:
                      type: string
                      enum: ['restart']
        404:
          description: Project with given projectID not found
        500:
          description: Internal error occurred
...
  /api/v1/projects/{id}/logs:
    get:
      summary: Retrieve the list of logs available for a project.
      description: >
        This requests the list of logs currently available for a project.
        This list may change over time as a project is built and run.
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      responses:
        200:
          description: Lists of the logs available grouped by type, build or app.
          content:
            application/json:
              schema:
                type: object
                properties:
                  build:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogDetails'
                  app:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogDetails'
        404:
          description: if the project with id was not found
    post:
      summary: Start streaming all logs for a given project
      description: >
        This requests that all logs currently available for a project start streaming.
        Logs are sent over socket.io via the 'log-update' event. When a log is enabled
        for streaming it's current content is sent in one or more 'log-update' events and
        then any further updates are sent via further 'log-update' events as they occur.
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
      responses:
        200:
          description: Lists of the logs that have been enabled grouped by type, build or app.
          content:
            application/json:
              schema:
                type: object
                properties:
                  build:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogDetails'
                  app:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogDetails'
        404:
          description: if the project with id was not found
        500:
          description: if an error occurred
    delete:
      summary: Stop streaming any enabled logs for a project.
      description: >
        This stops all current log streams for a project.
        If the log streams are re-enabled later then the logs will be sent from the beginning
        again.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: id of project
      responses:
        200:
          description: if the log streams were successfully stopped
        404:
          description: if the project with id was not found
        500:
          description: if an error occurred
...
  /api/v1/projects/{id}/logs/{type}/{name}:
    post:
      summary: Start streaming one log for a given project
      description: >
        This requests that the specified log for a project starts streaming.
        Logs are sent over socket.io via the 'log-update' event. When a log is enabled
        for streaming its current content is sent in one or more 'log-update' events and
        then any further updates are sent via further 'log-update' events as they occur.
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
        - in: path
          name: type
          schema:
            type: string
            enum: ['app', 'build']
          required: true
          description: type of log
        - in: path
          name: name
          schema:
            type: string
          required: true
          description: name of log
      responses:
        200:
          description: if the log stream was successfully enabled
        404:
          description: if the project or the specified log were not found
        500:
          description: if an error occurred
    delete:
      summary: Stop streaming one log for a given project
      description: >
        This requests that streaming stops for the specified log.
        If the log stream is re-enabled later then the log will be sent from the beginning.
        again.
      parameters:
        - in: path
          name: id
          schema:
            $ref: '#/components/schemas/ProjectID'
          required: true
          description: id of project
        - in: path
          name: type
          schema:
            type: string
            enum: ['app', 'build']
          required: true
          description: type of log
        - in: path
          name: name
          schema:
            type: string
          required: true
          description: name of log
      responses:
        200:
          description: if the log stream was successfully stopped
        404:
          description: if the project with id was not found
        500:
          description: if an error occurred
...
  /api/v1/projects/{id}/file-changes/{projectWatchStateId}/status:
    put:
      summary: Report the project watch state, either success or failed,  for a particular projectWatchStateId
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
        - name: projectWatchStateId
          in: path
          schema:
            type: string
          required: true
          description: projectWatchStateId represents a file watching client's current state. projectWatchStateId is generated and returned when first time calling the endpoint `GET /api/v1/projects/watchlist`. Additionally, if the ignored paths list for a project is changed, a new projectWatchStateId is generated and sent to file watchers via WebSocket event from `/websockets/file-changes/v1`
        - name: clientUUID
          in: query
          schema:
            type: string
            format: uuid
          required: true
          description: a unique id of the client
      requestBody:
        description: A JSON object to report the project watch state, either success or failed
        content:
          application/json:
            schema:
              type: object
              required:
                - success
              properties:
                success:
                  type: boolean
                  example: true
      
      responses:
        200:
          description: if successfully received the request
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /api/v1/projects/:id/file-changes:
    post:
      summary: Notify the event occurrance with a list of changed files of the project
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
        - name: timestamp
          in: query
          schema:
            type: number
          required: true
          description: timestamp of this file change event
        - name: chunk
          in: query
          schema:
            type: number
          required: true
          description: chunk number of this current chunk
        - name: chunk_total
          in: query
          schema:
            type: number
          required: true
          description: total chunks expected for this timestamp
      requestBody:
        description: A JSON array of which file paths have changed, which is then compressed with zlib, then converted to base64. The example shows JSON content before converted to base64
        content:
          application/json:
            schema:
              type: object
              required:
                - msg
              properties:
                msg:
                  type: array
                  items:
                    $ref: '#/components/schemas/FileChangeList'
      responses:
        200:
          description: if successfully received the request
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
...
components:
  schemas:
...
    ProjectID:
      type: string
      format: uuid
      example: d290f1ee-6c54-4b01-90e6-d701748f0851
      pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
...
    LogDetails:
      type: object
      required:
        - logName
      properties:
        logName:
          type: string
          example: 'docker.build.log'
        workspaceLogPath:
          type: path
...
    FileChangeList: 
        type: object
        required: 
          - path
          - timestamp
          - type
          - directory
        properties:
          path:
            description: the absolute path on the file system to the file
            type: string
            example: Documents/projectDir/File1
          timestamp:
            description: the timestamp of this change
            type: number
            example: 2117514496
          type:
            description: the change type, one of the following CREATE / MODIFY / DELETE
            type: string
            example: CREATE
          directory:
            description: true or false, on whether the path refers to a directory
            type: boolean
            example: false
...
  responses:
    400:
      description: Bad request
      content:
        application/json:
          schema:
            type: string
    404:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            type: string
    500:
      description: Internal error occurred.
      content:
        application/json:
          schema:
            type: string